feat: Optimize the structure for code readability
