+++ docs/images/installation.png
