+++ docs/images/oauth1.png
